# Conflicts:
#	.idea/modules.xml
#	.idea/workspace.xml

+++ Assets/Custom/Script/CustomerManagement.js
@@ -1,6 +1,11 @@
+let selectedCustomer = sessionStorage.getItem("selectedCustomer");
+
 var branchID;
+var customers = [];
+
 $(document).ready(function ()
 {
+    GetRequest("common/rolesAndBranches", customerSuccess);
     var branch =  JSON.parse(sessionStorage.getItem("branch"));
    // console.log(branch);
     branchID = branch.id;
@@ -17,7 +22,7 @@ function GetCustomers()
 //
 //}
 
-//function SuccessFarmerGet(Response)
+
 //{
 //    console.log(Response);
 //}
@@ -28,13 +33,13 @@ function SuccessCustomerGet(Response)
     console.log(Response);
     for(let count = 0; count < Response.length; count++){
         $('#customerTable').append('<tr>'+
-                                          '<td scope="row">'+ "F0" + Response[count].id +'</td>'+
+                                          '<td scope="row">'+ "C0" + Response[count].id +'</td>'+
                                           '<td>'+Response[count].name+'</td>'+
                                           '<td>'+ Response[count].address +'</td>'+
                                           '<td>'+ Response[count].branch.address+'</td>'+
                                           '<td>'+
                                           '<a  href="javascript:void(0);" data-toggle="modal" data-target="#VerifyModal">'+
-                                                                                           '<button  type="submit" class="btn btn-sm btn-success w-75">'+
+                                                                                           '<button  type="submit" class="btn btn-sm btn-success w-75" onClick="UpdateCustomer('+ Response[count].id +')">'+
                                                                                                 'Edit'+
                                                                                             '</button>'+
                                                                                         '</a>'+
@@ -50,11 +55,27 @@ function SuccessCustomerGet(Response)
                                       '</tr>')
 
     }
+
+    customers = Response;
+    if (Response.length > 0 && selectedCustomer != undefined) {
+        let Index = customers.findIndex(Cus => Cus.id === parseInt(selectedCustomer));
+        if (Index > -1) {
+            $('#inputName').val(customers[Index].name);
+            $('#address').val(customers[Index].address);
+        }
+    }
+
 }
 
 function DeleteCustomer(customerId) {
 
-    DeleteRequest("customer/"+customerId, {},SuccessCustomerDelete);
+
+    if (confirm("Are you sure to delete this farmer with ID C0" + customerId + "?")) {
+
+        DeleteRequest("customer/"+customerId, {},SuccessCustomerDelete);
+
+    }
+
 
 }
 
@@ -62,8 +83,117 @@ function SuccessCustomerDelete()
 {
     console.log("Deleted");
     GetCustomers();
+}
+
+
+//Update Functions
+
+function UpdateCustomer(customerId) {
+
+    sessionStorage.setItem("selectedCustomer", customerId);
+    window.location.href = "EditCustomer.html";
 
 }
+
+function customerSuccess(response) {
+    for (const branch of response.branches) {
+        $('#selectBranch').append(new Option(branch.address, branch.id));
+    }
+
+    $('#selectBranch').val(branchID);
+}
+
+
+$('#updateCustomer').click(function () {
+    clearValidations();
+    var formData = $('#formUpdate').serializeObject();
+    var isAnyError = updateValidation(formData);
+
+    if (!isAnyError) {
+        let requestData = setRequestData(formData);
+
+        requestData.branch.id = branchID;
+        requestData.id = selectedCustomer;
+        console.log(requestData);
+        UpdateRequest("customer", requestData, updateCustomerSuccess);
+
+    }
+});
+
+$('#cancelUpdate').click(function () {
+
+        window.location.href = "CustomerManagement.html";
+
+});
+
+var updateValidation = function (formData) {
+
+    var errorText = "";
+    var isAnyError = false;
+
+    if (formData.inputName == "") {
+        errorText = "Please Provide a Name"
+        isAnyError = setError("inputName");
+
+    } else if (formData.address == "") {
+        errorText = "Please Provide a Address"
+        isAnyError = setError("address");
+
+    } else if (formData.branch == "") {
+        errorText = "Please select a Branch"
+        isAnyError = setError("selectBranch");
+    }
+
+    if (isAnyError) {
+        $('#errorAlert').text(errorText);
+        $('#errorAlert').removeClass("d-none");
+    }
+
+    return isAnyError;
+}
+
+var setError = function (element) {
+    $("#" + element).addClass("border-danger");
+    $("#" + element).focus();
+    return true;
+}
+
+var clearValidations = function () {
+
+    $('#inputName').removeClass("border-danger");
+    $('#address').removeClass("border-danger");
+    $('#selectBranch').removeClass("border-danger");
+    $('#errorAlert').text("");
+    $('#errorAlert').addClass("d-none");
+}
+
+var setRequestData = function (formData) {
+
+    var customer = {
+        name: formData.inputName,
+        address: formData.address,
+        branch: {
+            id: formData.branch
+        }
+    }
+
+    return customer;
+}
+
+function updateCustomerSuccess(response) {
+    sessionStorage.setItem("branch", JSON.stringify(response.data.branch));
+
+    PopUpWithTitleAndText("Success","Customer Details Updated","success");
+
+    setTimeout(function()
+    {
+        window.location.href = "CustomerManagement.html";
+    },2000);
+
+}
+
+
+
